LayoutTests/fast/attachment/attachment-dom-expected.txt

This tests that attachments have an instance type of HTMLElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS attachment is an instance of HTMLAttachmentElement
PASS attachment.file is null
Setting attachment.file to a file created from resources/test-file.txt
PASS attachment.file is file
PASS attachment.file.name is 'test-file.txt'
PASS successfullyParsed is true

TEST COMPLETE
